Rot(90), dark theme, lock
The map has been rotated 90 degrees, to display it vertically.
A dark mode has been added.
A lock is available on the application, the keys are the 4 pieces of furniture on the top right of the workshop. The lock can give access to hidden pages or lock and unlock certain buttons.